requirements.txt
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision ==0.11.3+cu113
PyQt5==5.15.4
numpy==1.19.5
dlib==19.7.0
opencv-python==4.5.5.64
Pillow==8.4.0
scikit-learn==0.24.2
ninja==1.11.1.1